various fixes
Add a Dockerfile.development to create a docker image that has prebuilt all kmods.
the idea for now is to have a Dockerfile that create an image as small as possible.
and another more complete with all kmods.

Note: the removal of git history for openwrt feeds will be removed later because potentially unsafe.

Swqetup a build script that does a first build with '-i' or '--ignore-errors'
on main make command and then save a config file with profiles to little for the
specified packages set.